% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwsdar_functions.R
\name{expit}
\alias{expit}
\title{Exponential Inverse Logit (Expit) Function}
\usage{
expit(x)
}
\arguments{
\item{x}{Input numeric vector}
}
\value{
Transformed values in (0,1) range
}
\description{
Computes the inverse logit (expit) transformation: expit(x) = 1 / (1 + exp(-x))
}
\examples{
expit(c(-1, 0, 1))

}
